[VARKA] Divide through double lanes at the 64-bit lane
Task 88 step 2 landed the double route at the int lane, where a division
converts into two double vectors of half the lanes each, divides, converts back
and rejoins the halves with an `or` - seven operations. At the 64-bit lane a
lane and a double lane are the same width, so `part` is 0 both ways, there is no
second half and nothing to rejoin: one L2D, one divide, one D2L. Three
operations, and no magic alternative to weigh them against, since no 64-bit
multiply-high exists.

`ConstDivide` therefore accepts a LONG child and carries a `long` divisor, which
the TIME divisors need - nanos per hour is 3.6e12. Two refusals come with it: an
int-lane divisor that does not fit in an int, whose every quotient would be
zero, and the existing -1 refusal, which now names the lane's own most negative
value.

The 64-bit lane is exact only under a bound and this node cannot check one it is
not handed, so `EXACT_DIVIDEND_BOUND` states it - 2^53, a bound on the dividend
that does not depend on the divisor, derived by
`sql/varka/plans/verify_double_division.py` - and the caller owns proving it.
For TIME the bound is structural: nanoseconds of day are under 2^47.

Tests: the seven divisors tasks 102 and 103 need, each over dividends derived
from its own divisor rather than written out - both signs of d, d-1, d+1, 2d and
both ends of the exact range - at both long widths and every null pattern,
against the reference evaluator; the op counts, one divide and two conversions
against the int lane's two and four; and the refusals. `VarkaLaneTypeSuite`'s
long-lane table gains the node, which is what makes the widening visible to a
reader rather than only to the emitter.

Generated-by: Claude Code (Claude Opus 5)

+++ sql/catalyst/src/test/scala/org/apache/spark/sql/catalyst/expressions/codegen/varka/VarkaReferenceEvaluator.scala
@@ -150,7 +150,10 @@ object VarkaReferenceEvaluator {
     // rather than reproducing the emitter's conversion, so the two agree only if the lowering
     // is right.
     case n: ConstDivide =>
-      evalValue(n.child(), row, lits).map(_ / n.divisor())
+      // The divisor is declared at the wider lane and narrowed here, which the node's own
+      // constructor makes safe: an int-lane division whose divisor does not fit in an int is
+      // refused where it is built.
+      evalValue(n.child(), row, lits).map(_ / n.divisor().toInt)
     case n: MakeDate =>
       // The definition: LocalDate.of, null (None) where the calendar rejects the triple - never
       // the length rule the emitter computes. The year limit is the kernel's business, not the
